fix(spans): moving AnnotatedValue to _types to resolve circular import
